@@ -1,4 +1,4 @@
-package cn.luischen;
+package cn.luischen.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
