@@ -1,7 +1,7 @@
 package cn.luischen.utils;
 
-import sun.misc.BASE64Decoder;
-import sun.misc.BASE64Encoder;
+import java.util.Arrays;
+import java.util.Base64;
 
 import javax.crypto.Cipher;
 import javax.crypto.spec.SecretKeySpec;
@@ -59,14 +59,14 @@ public class Tools {
         SecretKeySpec secretKeySpec = new SecretKeySpec(key.getBytes("UTF-8"), "AES");
         cipher.init(Cipher.ENCRYPT_MODE, secretKeySpec);
         byte[] encryptedBytes = cipher.doFinal(data.getBytes());
-        return new BASE64Encoder().encode(encryptedBytes);
+        return Arrays.toString(Base64.getMimeEncoder().encode(encryptedBytes));
     }
 
     public static String deAes(String data, String key) throws Exception {
         Cipher cipher = Cipher.getInstance("AES");
         SecretKeySpec secretKeySpec = new SecretKeySpec(key.getBytes("UTF-8"), "AES");
         cipher.init(Cipher.DECRYPT_MODE, secretKeySpec);
-        byte[] cipherTextBytes = new BASE64Decoder().decodeBuffer(data);
+        byte[] cipherTextBytes = Base64.getDecoder().decode(data);
         byte[] decValue = cipher.doFinal(cipherTextBytes);
         return new String(decValue);
     }
